requirements.txt
opencv-python
numpy
natsort
matplotlib
tqdm
scikit-learn
pillow